AB-1412:fix:Netlify problem due to asset import, fixing by moving asset folder and reimporting
